requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
fpdf==1.7.2
idna==3.4
Pillow==9.5.0
requests==2.28.2
urllib3==1.26.15
